Changes the default configuration for the KeyZones Module and reduces the volume state duration from 5 to 3 minutes
The KeyZones configuration has been altered so the chances of missing a contact are reduced. The new build makes use of 10k candlestick items (~104 days), increasing the number of zones significantly. The volume duration has been reduced as 5 minutes seems to be a bit long considering the state test is executed every minute.